Convert float from f32 to f64
With f32 there where not enough numerical precision with some
computation and the final results was infinite.

+++ src/lib.rs
@@ -18,14 +18,14 @@ fn soft_dtw_rust(m: &Bound<'_, PyModule>) -> PyResult<()> {
 // Function in the python module
 
 // #[pyfunction] // NOT WORK
-fn get_distance_matrix(x : PyReadonlyArrayDyn<f32>, y : PyReadonlyArrayDyn<f32>, gamma : Option<f32>) -> PyResult<Array2<f32>> {
+fn get_distance_matrix(x : PyReadonlyArrayDyn<f64>, y : PyReadonlyArrayDyn<f64>, gamma : Option<f64>) -> PyResult<Array2<f64>> {
     // sdtw_v1::print_type_of(&x);
 
     Ok(sdtw_v1::compute_distance_matrix(x.as_array(), y.as_array(), gamma.unwrap_or(1.)))
 }
 
 #[pyfunction]
-fn compute_sdtw_1d(x : PyReadonlyArrayDyn<f32>, y : PyReadonlyArrayDyn<f32>, gamma : Option<f32>) -> PyResult<f32> {
+fn compute_sdtw_1d(x : PyReadonlyArrayDyn<f64>, y : PyReadonlyArrayDyn<f64>, gamma : Option<f64>) -> PyResult<f64> {
     // sdtw_v1::print_type_of(&x);
     Ok(sdtw_v1::compute_sdtw(x.as_array(), y.as_array(), gamma.unwrap_or(1.)))
 }
